[refactor] Modify class name for clear naming
